+++ packages/react-article-components/src/components/related/mockup.js
@@ -1,0 +1,44 @@
+export default {
+  mobile: {
+    thumbnail: {
+      width: 88, // px
+      height: 68, // px
+    },
+    item: {
+      width: 'calc((309/355)*100%)',
+    },
+  },
+  tablet: {
+    thumbnail: {
+      width: 88, // px
+      height: 68, // px
+    },
+    item: {
+      width: 513, // px
+    },
+  },
+  desktop: {
+    thumbnail: {
+      width: 246, // px
+      height: 148, // px
+    },
+    item: {
+      margin: {
+        right: 8, // px
+      },
+      width: 246, // px
+    },
+  },
+  hd: {
+    thumbnail: {
+      width: 349, // px
+      height: 148, // px
+    },
+    item: {
+      margin: {
+        right: 10, // px
+      },
+      width: 349, // px
+    },
+  },
+}